🚚 move settings forms in their own file
