+++ frontend-new/app/routes/services/settings/service-ports-form.tsx
@@ -1,0 +1,44 @@
+import { ExternalLinkIcon, TriangleAlertIcon } from "lucide-react";
+import { Code } from "~/components/code";
+import { Alert, AlertDescription, AlertTitle } from "~/components/ui/alert";
+
+export type ServicePortsFormProps = {
+  project_slug: string;
+  service_slug: string;
+};
+
+export function ServicePortsForm({
+  service_slug,
+  project_slug
+}: ServicePortsFormProps) {
+  return (
+    <div className="w-full max-w-4xl flex flex-col gap-5">
+      <div className="flex flex-col gap-3">
+        <h3 className="text-lg">Exposed ports</h3>
+        <p className="text-gray-400">
+          This makes the service reachable externally via the ports defined
+          in&nbsp;
+          <Code>host port</Code>. Using&nbsp;
+          <Code>80</Code>
+          &nbsp;or&nbsp;
+          <Code>443</Code>
+          &nbsp;will create a default URL for the service.
+        </p>
+
+        <Alert variant="warning">
+          <TriangleAlertIcon size={15} />
+          <AlertTitle>Warning</AlertTitle>
+          <AlertDescription>
+            Using a host value other than 80 or 443 will disable&nbsp;
+            <a href="#" className="underline inline-flex gap-1 items-center">
+              zero-downtime deployments <ExternalLinkIcon size={12} />
+            </a>
+            .
+          </AlertDescription>
+        </Alert>
+      </div>
+    </div>
+  );
+}
+
+function NewServicePortForm() {}
